fix New Feed Selection
In my tests the way I had written it first time around did always work, and given the data is pulled from a live URL I have a hard time believing it was just luck that the feedloads were executing before the js had finished. Probbly it was because my entire feedreader spec file is wrapped in the jQuery function so isn't running until DOM is loaded.

Should now be fixed to properly wait.

Signed-off-by: ChrisBoon <cboon27@gmail.com>